NNS1-3480: Refactors mapIcpTransaction to allow for the introduction of a more generic function (#5891)
# Motivation

Transactions need to be adapted to a user-friendly format. We already
have a function called `mapIcpTransaction` that assists with this, but
it is too specific to its current use case.
We can split it into two functions: a generic one that resolves
transaction information and a specific one for the transactions table.
This first PR introduces changes that will allow for the implementation
of a more generic function in a second PR, which will be reused in a
third PR.

# Changes

- Renames `mapIcpTransaction` to `mapIcpTransactionToUi`.  
- Moves the logic for differentiating between transaction types `Send`
and `Receive` from `transactionName` to `getTransactionType` in both
`icp-transactions` and `icrc-transactions`.
-  Introduces a new type for `Receive` in `AccountTransactionType`.

# Tests

- Removes the unnecessary test for `transactionName` that checks for
differentiation between `Send` and `Receive`.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.