Remove timestamp from icpSwapTickersStore (#5901)
# Motivation

I thought there was a requirement for the ICP Swap exchange rates to be
updated every 10 minutes.
So I added a timestamp in the store to see when they were last updated.
But this turned out to be a misunderstanding and we want the exchanges
rates to remain fixed until the session ends (within 30 minutes).
So there is no need to have a timestamp in the store.

# Changes

Remove `lastUpdateTimestampSeconds` from `icpSwapTickersStore` and make
the array of tickers the stored data.

# Tests

Updated

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary