Add ledgerCanisterId to UserTokenBase (#5905)
# Motivation

ICP Swap provides exchange rates per token based on ledger canister ID.
So in the tokens table, to get the exchange rate of a token, we need to
know its ledger canister ID.
`UserTokenBase` is the base type of row data used in the tokens table.

In this PR we just make sure that the ledger canister ID is available
for each token in the table.
This can then be used in later PRs to get the exchange rate.

# Changes

1. Add `ledgerCanisterId` field to `UserTokenBase`.
2. When constructing instances of `UserTokenBase`, add
`ledgerCanisterId`.

# Tests

1. Unit tests have been updated.
2. Tested manually in another branch that we can display dollar values
in the tokens table.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet