Refactor neuron tag (#5908)
# Motivation

For inactive neurons, a 🔴 red tag (styled as `danger`) needs to be
shown. Currently, in many places, the information for neuron tags is
provided as a `string` and rendered using the base `Tag` component. With
this PR, we replace the `string` with `NeuronTagData` for tag data type
and replace the `Tag` component with the `NeuronTag` component.

Note: The neuron tag will be extended with the status in a separate PR.

# Changes

- Add size property to the NeuronTag.
- Replace string with NeuronTagData for the neuron tag data.
- Replace Tag with NeuronTag component for neuron tags.

# Tests

- Added NeuronTagPo
- Updated tests to use it.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary