Fix: Adds an override for the name and token.symbol values for abandoned SNS (#5915)
# Motivation

As discussed
[here](https://dfinity.slack.com/archives/C03H6QEPW5D/p1733263773416409)
and
[here](https://dfinity.slack.com/archives/C039M7YS6F6/p1733302975333649),
we have a situation with the SNS `CYCLES-TRANSFER-STATION`. They have
decided to change all their metadata, resulting in the name and token
being displayed as `---`.

This first PR introduces a method to override the values we receive from
the SNS-Aggregator, allowing the rest of the application to use them.

# Changes

- Checks a dictionary of "broken" SNSs for matches and overrides `name`
and `token.symbol` for any that are found.

# Tests

- Unit test for the aggregator store derived data to verify that when
`rootCanisterId` is present in the dictionary, the values for `name` and
`token.symbol` are replaced.

# Todos

- [ ] Add entry to changelog (if necessary).