NNS1-3497: new function to sort SNS projects from newest to oldest (#5948)
# Motivation

This is a follow-up to #5918 regarding the display of SNS projects in
the Launchpad section. We aim to sort them in descending order, from
newest to oldest.
This initial PR introduces a function to sort based on
`swapOpenTimestampSeconds`.
The subsequent PR will implement the requirements outlined in
[NNS-3497](https://dfinity.atlassian.net/jira/software/c/projects/NNS1/boards/333?selectedIssue=NNS1-3497).

# Changes

- New descending sort function based on SNS summary
`swapOpenTimestampSeconds`.

# Tests

- Unit tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary