Show error state when ICPSwap API call fails. (#5960)
# Motivation

We loaded token prices from ICP Swap.
If this results in an error, we want to show `-/-` as USD values and
show an error state with tooltip.
There can be 2 reasons why there is an error:
1. Fetching the ICP Swap data fails.
2. Parsing the ICP Swap data fails.

In this PR we set the error state when fetching the data fails.

# Changes

1. Add `"error"` as a possible value for `initIcpSwapTickersStore`.
2. Forward `"error"` from `initIcpSwapTickersStore` to
`icpSwapUsdPricesStore` (which already has the state for when parsing
fails).
3. Set `"error"` when the API call fails in the service.

# Tests

1. Unit tests added.
2. Added manually by adding `throw new Error()` in the API function.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet