Show tooltip when price data is missing for some tokens (#5958)
# Motivation

At the top of the tokens page, we want to show the total value of the
user's tokens.
If we don't have price data for some tokens, we treat these tokens as
having no value.
To indicate this, we show an extra tooltip if the user owns any tokens
which we ignored in computing the total.

# Changes

1. Add `hasUnpricedTokens` prop to `UsdValueBanner`.
2. If the prop is true, show an extra tooltip.
3. Pass `hasUnpricedTokens` from `Tokens.svelte` based on whether the
user has any tokens with positive `balance` but undefined
`balanceInUsd`.

# Tests

1. Unit tests added.
2. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [ ] Add entry to changelog (if necessary).
not yet