Missing rewards soon tag (#5961)
# Motivation

Neurons that are about to start missing rewards need to be marked with a
tag to draw users' attention, indicating that action needs to be taken
to avoid missing rewards. This PR introduces a "warning" state to the
`NeuronsTag` component and applies it to the tags for "missing soon"
neurons.

# Changes

- Added a "warning" neuron tag status.
- Introduced a "...to confirm" tag.

# Tests

- Added.
- Tested manually.

<img width="1195" alt="image"
src="https://github.com/user-attachments/assets/1a105dd1-c64f-449c-bef4-0dd5fb57b703">


# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.