Replace nowInSeconds() with a static value (#5987)
# Motivation

One test is
[flaky](https://github.com/dfinity/nns-dapp/actions/runs/12276181573/job/34252843443)
because the mocked data was created during test initialization with the
current time value, assuming it wouldn't change. However, the system
time was updated before every test with the current time using
`vi.useFakeTimers().setSystemTime(Date.now())`. This results in a failed
test due to date rounding in `nowInSeconds`.

# Changes

- Use the same time for both mock data calculation and system time.

# Tests

- Pass.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.