Reset following neuron state (#5998)
# Motivation

In the case where a neuron hasn't been used for a long period of time
(currently defined as 6 + 1 months), its followings are reset. To avoid
confusion, it was decided to display a special type of inactive status
with additional information in the description and a "Set Following"
button, instead of "Confirm Following"

**Note:** Currently, the component looks suboptimal with such a long
description. This issue will be addressed in a separate PR.

# Changes

- New `isNeuronFollowingReset` util.
- Display additional Neuron reward state.

# Tests

- Updated unit tests.
- Tested manually:
<img width="717" alt="image"
src="https://github.com/user-attachments/assets/3d345d8c-2359-4b5c-bbfa-461fe141e7ff"
/>


# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.