Add stakeInUsd field to TableNeuron (#6017)
# Motivation

We want to show USD values in the neurons tables, similar to the tokens
table and the projects table.

In this PR we just add the USD stake to the `TableNeuron` objects. It
will be displayed in future PRs.

# Changes

1. Add `stakeInUsd` field to `TableNeuron` type.
2. Populate `stakeInUsd` in and `tableNeuronsFromNeuronInfos` and
`tableNeuronsFromSnsNeurons`.
3. Make `icpSwapUsdPrices` temporarily an optional parameter until we
pass it from all the components callers in future PRs. Then they will be
tested too.

# Tests

1. Unit tests adapted.
2. Tested manually (including future changes) at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not yet