Support rendering stake in USD in NeuronsTable (#6018)
# Motivation

We want to show USD values in the neurons tables, similar to the tokens
table and the projects table.

This PR adds support for rendering the `stakeInUsd` field in the
`NeuronsTable` components.

Still missing (for future PRs):
1. Passing `icpSwapUsdPrices` to `tableNeuronsFromNeuronInfos` and
`tableNeuronsFromSnsNeurons`.
2. Calling `loadIcpSwapTickers` on the neurons table page to make sure
data is available.

# Changes

1. Render stake in USD in `NeuronStakeCell.svelte` if
`ENABLE_USD_VALUES_FOR_NEURONS` is enabled.

# Tests

1. Unit tests added.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not yet