Pass icpSwapUsdPricesStore when constructing table neurons (#6027)
# Motivation

In order to display neuron stakes in USD we need token price data to
factor into the calculation.

# Changes

1. Make `icpSwapUsdPrices` non-optional parameter to
`tableNeuronsFromNeuronInfos` and `tableNeuronsFromSnsNeurons`. It can
still be `undefined` but should only be so when the ICPSwap data isn't
loaded yet.
2. Pass `icpSwapUsdPrices` from `NnsNeurons` and `SnsNeurons` components
to display in the tables.
3. Include `icpSwapUsdPrices` in `tableNeuronsToSortStore` and
`snsTableNeuronsToSortStore`. This currently has no effect but might if
we ever include `stakeInUsd` in the ordering logic. It's good to have
for completeness.

# Tests

1. Unit tests added for `NnsNeurons` and `SnsNeurons`.
2. Since the USD price currently has no effect on ordering, there's
nothing to test for `neuronsTableOrderSortedNeuronIdsStore` or
`snsNeuronsTableOrderSortedNeuronIdsStore`.
3. Tested manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not yet