Extract HeadingSubtitleWithUsdValue (#6037)
# Motivation

On the wallet page, we show the account balance in USD in the heading
subtitle.
We want to do something similar on the NNS and SNS neuron details page.

To make this easier, in this PR we extract a component to show a heading
subtitle with USD value.

# Changes

1. Extract `HeadingSubtitleWithUsdValue` from `WalletPageHeading`.
2. Use `HeadingSubtitleWithUsdValue` in `WalletPageHeading` if
`ENABLE_USD_VALUES` is enabled.

# Tests

1. Unit test added.
2. Page objects adjusted.
3. Tested manually.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary