Show stake in USD on NNS neuron details (#6039)
# Motivation

Want to show the stake in USD on the neuron details page similar to how
we show the account balance on the wallet page.

This PR does it for NNS neurons. Another PR will do SNS neurons.
And a third PR will make sure token prices are loaded when the neuron
page is deep linked.

# Changes

1. Use `HeadingSubtitleWithUsdValue` which was extracted in
https://github.com/dfinity/nns-dapp/pull/6037 in `NnsNeuronPageHeading`.
Some of the slot content is temporarily duplicate between with and
without feature flag.

# Tests

1. Unit tests added.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neuron/?u=qsgjb-riaaa-aaaaa-aaaga-cai&neuron=7894867686608507388

# Todos

- [ ] Add entry to changelog (if necessary).
not yet