refactor: UiTransactionsList rendering logic (#6048)
# Motivation

We fixed a minor bug in https://github.com/dfinity/nns-dapp/pull/5874
that prevented the `completed` property from being set to `true` for
empty transactions. The `UiTransactionsList` uses this property to
disable the list once we have fetched all the items. However, it renders
an Empty component when there are no transactions, which obscured the
bug.

This PR aims to simplify the logic and enhance readability.

# Changes

* Refactor code to simplify the logic deciding what to render in the
`UiTransactionsList`

# Tests

- Added a new test to check the `Spinner`.  
- Added checks to existing tests for UI elements.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary