Display missing rewards indicator in navigation (#6071)
# Motivation

When the user has Nns neurons that haven’t been active and are set to
start losing rewards in 30 days or less, a red circle needs to be
displayed next to the “Neuron Staking” sidebar item and on in the Neuron
Staking page table to indicate that action is required.
In this PR, we add the indicator component next to the “Neuron Staking”
navigation entry.

# Changes

- Display the indicator in the menu.

# Tests

- Test the presence of the indicator component. The logic is tested in
the `NnsNeuronsMissingRewardsBadge.specs` file.

## Screenshot

<img width="350" alt="image"
src="https://github.com/user-attachments/assets/42445884-4852-44d4-9d0c-0b51f6b82fce"
/>



# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.