Display missing rewards indicator in projects table (#6072)
# Motivation

When the user has Nns neurons that haven’t been active and are set to
start losing rewards in 30 days or less, a red circle needs to be
displayed next to the “Neuron Staking” sidebar item and on in the Neuron
Staking page table to indicate that action is required.
In this PR, we add the indicator component in the projects table.

# Changes

- Display the indicator in the table.

# Tests

- Test the presence of the indicator component. The logic is tested in
the NnsNeuronsMissingRewardsBadge.specs file.

## Screenshot
<img width="925" alt="image"
src="https://github.com/user-attachments/assets/d70b3e60-a08b-4427-aed2-4a1b5447dd32"
/>

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary