Skip download test if last commit is too old (#6091)
# Motivation

We keep build artifacts for 3 days as specified
[here](https://github.com/dfinity/nns-dapp/blob/b8e278aa4361df2abe0cd9cc7dda69b591cb73e5/.github/workflows/build.yml#L52).
We have a script for downloading build artifacts, which we test for the
main branch.
This test fails if artifacts for the latest commit to the main branch
have already expired, as can be seen
[here](https://github.com/dfinity/nns-dapp/actions/runs/12592378031/job/35096899373?pr=6089).

# Changes

Skip the test if the latest commit is too old.
This should be rare so hopefully on the next commit the test will run
again.

# Tests

The test passes with
```
```

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary