Refactor attach_canister to prepare for block_index (#6104)
# Motivation

To support moving the fallback logic for canister creation to the
frontend, we want to add a `block_index` field to the `NamedCanister`
type stored in the canister.
See [this
doc](https://docs.google.com/document/d/1hjMSTzjnVbU9Q4rJk233M3uNcKC-RLhme25nahHpTZg/edit?tab=t.0#heading=h.qtiuv2x3gsjy)
for details.

In this PR we don't yet add the new field but change the logic to make
it easier to add a new field.

# Changes

1. Change the code in `attach_canister` to be more generic so it
requires fewer changes when we add the `block_index` field to
`NamedCanister`.
2. Change `rename_canister` so it changes only the name and keeps
everything else from the canister. Currently "everything else" is just
the canister ID but this is more resilient when we add more fields.

# Tests

1. New unit tests were added to expand test coverage. Tests were added
before the code was changed to make sure the behavior was not changed in
the refactoring.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary