NNS1-3527: Allow dfinity: prefix for ICP payments (#6108)
# Motivation

Coinbase incorrectly uses `dfinity:` instead of `icp:` as the token
symbol in the QR payment URI. We've asked them to fix this but so far
they haven't. So we work around this by allowing `dfinity:` in place of
`icp:`.

# Changes

1. Don't reject the payment URL from the QR code if the prefix is
`dfinity:` but the required token is ICP.

# Tests

1. Added a `MockQRCodeReaderModal` component which has just a text input
to fake what was scanned by the QR code reader.
2. Added missing unit tests using the `MockQRCodeReaderModal`.
3. Add unit tests for `dfinity:` prefix.

# Todos

- [x] Add entry to changelog (if necessary).