Fix errors logged in SnsProposalDetail.spec.ts (#6120)
# Motivation

These 2 tests in `SnsProposalDetail.spec.ts`:
```
"should navigate to the proposal from the previous Sns"
"should navigate to the proposal from the next Sns"
```
are currently logging errors like this:
```
{
  certified: false,
  strategy: 'query_and_update',
  error: t [Error]: No proposal for given proposalId 19
      at Object.queryProposal (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/tests/fakes/sns-governance-api.fake.ts:327:11)
      at fn (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/tests/utils/module.test-utils.ts:158:33)
      at wrapMaybePaused (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/tests/utils/module.test-utils.ts:140:20)
      at Module.pausableFunctions.<computed> (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/tests/utils/module.test-utils.ts:157:14)
      at Module.mockCall (file:///Users/dskloet/dev/nns-dapp/tree4/frontend/node_modules/@vitest/spy/dist/index.js:61:17)
      at Module.queryProposal (file:///Users/dskloet/dev/nns-dapp/tree4/frontend/node_modules/tinyspy/dist/index.js:45:80)
      at request (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/lib/services/public/sns-proposals.services.ts:209:7)
      at queryOrUpdate (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/lib/services/utils.services.ts:85:5)
      at Module.queryAndUpdate (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/lib/services/utils.services.ts:105:17)
      at Module.getSnsProposalById (/Users/dskloet/dev/nns-dapp/tree4/frontend/src/lib/services/public/sns-proposals.services.ts:206:10),
  identity: {
    getPrincipal: [Function: getPrincipal],
    transformRequest: [Function: transformRequest]
  }
}
```

This started happening after
https://github.com/dfinity/nns-dapp/pull/5788
Before that PR, the tests were relying on `console.error` being mocked
in the `beforeEach` of a different `describe` block but after that PR
all mocks are always restored before each test so the mocking of
`console.error` was being reverted.

However logging during tests is not allowed and normally results in test
failure. However these test passed before the logging happened so the
logging could happen without the tests failing.

And in general mocking `console.error` in `beforeEach` is dangerous
because it can hide problems in other tests.
It's best to only mock `console.error` in tests that expect logging and
always expect the specific messages being logged so no other errors are
hidden.

# Changes

1. Stop mocking `console.error` in `beforeEach` in
`SnsProposalDetail.spec.ts`.
2. Mock `console.error` in specific tests where errors are logged and
expect what is being logged.
3. In `"show neurons that can vote"`, add a missing proposal that the
test expects. It was logging errors for this that were ignored. The test
does not seem to intend the proposal to be missing.
4. Add `await runResolvedPromises();` in `afterEach` in
`failTestsThatLogToConsole` in order to prevent tests from passing
before they log some errors.
5. In `queryProposal` in the fake SNS API, take into account that there
may not be any proposals for an identity+SNS combination and allow to
throw the intended error instead of `undefined.proposals` not existing.
This happened in one of the tests making the logged error message even
more confusing.

# Tests

Tests only

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary